Pull Request: Fixing bug reported in #353 by refactoring the code.
updates #353
closes #353
Signed off: Michael A. Jackson

+++ Source/Plugins/Sampling/SamplingFilters/SampleSurfaceMesh.cpp
@@ -262,7 +262,7 @@ void SampleSurfaceMesh::execute()
 
   // create a dynamic list array to hold face lists
   Int32Int32DynamicListArray::Pointer faceLists = Int32Int32DynamicListArray::New();
-  std::vector<size_t> linkCount(numFeatures, 0);
+  std::vector<int32_t> linkCount(numFeatures, 0);
 
   // fill out lists with number of references to cells
   typedef boost::shared_array<int32_t> SharedInt32Array_t;
